+++ drift/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add support for subqueries in the Dart query builder.
 - Add `isInExp` and `isNotInExp` to construct `IS IN` expressions with arbitrary
   expressions.
+- Add the `substr` extension on `Expression<String>` to call the sqlite3 function from
+  the Dart API.
 - Add `isolateSetup` to `NativeDatabase.createInBackground()` to override libraries.
 
 ## 2.10.0